Strip out bag2video_rospy, use rosbags only
