+++ pyproject.toml
@@ -33,8 +33,8 @@ classifiers = [
 Homepage = "https://github.com/apl-ocean-engineering/bag2video"
 
 [project.scripts]
-bag2images = "bag2video_rosbags.bag2images:main"
-bag2video = "bag2video_rosbags.bag2video:main"
+bag2images = "bag2video.bag2images:main"
+bag2video = "bag2video.bag2video:main"
 
 [tool.pdm]
 distribution = true
